change f64 to f32 for percentages

+++ accounting_system_rust/invoice_doc_generator/src/percentages/tax_discount_cess.rs
@@ -29,16 +29,16 @@ impl TryFrom<i32> for GSTPercentage {
 }
 
 #[derive(Debug)]
-pub struct TaxPercentage(f64);
+pub struct TaxPercentage(f32);
 
 #[derive(Debug, Error)]
 pub enum TaxPercentageError {
     #[error("tax percentage {0} cannot be less than 0 or greater than 100")]
-    NotInBounds(f64),
+    NotInBounds(f32),
 }
 
 impl TaxPercentage {
-    pub fn new(tax_percentage: f64) -> Result<Self, TaxPercentageError> {
+    pub fn new(tax_percentage: f32) -> Result<Self, TaxPercentageError> {
         if !(0.0..=100.00).contains(&tax_percentage) {
             return Err(NotInBounds(tax_percentage));
         }
@@ -60,7 +60,7 @@ mod tax_percentage_tests {
     #[case(101.0, false)]
     #[case(0.0, true)]
     #[case(20.0, true)]
-    fn test_tax_percentage(#[case] input: f64, #[case] valid: bool) {
+    fn test_tax_percentage(#[case] input: f32, #[case] valid: bool) {
         let p = TaxPercentage::new(input);
         if valid {
             assert_that!(p).is_ok();
@@ -71,30 +71,30 @@ mod tax_percentage_tests {
 }
 
 #[derive(Debug, Serialize, Deserialize,Clone)]
-#[serde(try_from = "f64")]
-pub struct CessPercentage(f64);
+#[serde(try_from = "f32")]
+pub struct CessPercentage(f32);
 
 #[derive(Debug, Error)]
 pub enum CessPercentageError {
     #[error("cess percentage {0} cannot be less than 0 or greater than 500")]
-    NotInBounds(f64),
+    NotInBounds(f32),
 }
 
 impl CessPercentage {
-    pub fn new(cess_percentage: f64) -> Result<Self, CessPercentageError> {
+    pub fn new(cess_percentage: f32) -> Result<Self, CessPercentageError> {
         if !(0.0..=500.00).contains(&cess_percentage) {
             return Err(CessPercentageError::NotInBounds(cess_percentage));
         }
         Ok(Self(cess_percentage))
     }
-    pub fn inner(&self)->f64{
+    pub fn inner(&self)->f32{
         self.0
     }
 }
 
-impl TryFrom<f64> for CessPercentage {
+impl TryFrom<f32> for CessPercentage {
     type Error = anyhow::Error;
-    fn try_from(value: f64) -> Result<Self, Self::Error> {
+    fn try_from(value: f32) -> Result<Self, Self::Error> {
         CessPercentage::new(value).context("")
     }
 }
@@ -114,7 +114,7 @@ mod cess_percentage_tests {
     #[case(101.0, true)]
     #[case(0.0, true)]
     #[case(20.0, true)]
-    fn test_cess_percentage(#[case] input: f64, #[case] valid: bool) {
+    fn test_cess_percentage(#[case] input: f32, #[case] valid: bool) {
         let p = CessPercentage::new(input);
         if valid {
             assert_that!(p).is_ok();
@@ -125,31 +125,31 @@ mod cess_percentage_tests {
 }
 
 #[derive(Debug, Serialize, Deserialize,Clone)]
-#[serde(try_from = "f64")]
-pub struct DiscountPercentage(f64);
+#[serde(try_from = "f32")]
+pub struct DiscountPercentage(f32);
 
 #[derive(Debug, Error)]
 pub enum DiscountPercentageError {
     #[error("discount percentage {0} cannot be less than 0 or greater than 100")]
-    NotInBounds(f64),
+    NotInBounds(f32),
 }
 
 impl DiscountPercentage {
-    pub fn new(discount_percentage: f64) -> Result<Self, DiscountPercentageError> {
+    pub fn new(discount_percentage: f32) -> Result<Self, DiscountPercentageError> {
         if !(0.0..=100.00).contains(&discount_percentage) {
             return Err(DiscountPercentageError::NotInBounds(discount_percentage));
         }
         Ok(DiscountPercentage(discount_percentage))
     }
 
-    pub fn inner(&self)->f64{
+    pub fn inner(&self)->f32{
         self.0
     }
 }
 
-impl TryFrom<f64> for DiscountPercentage {
+impl TryFrom<f32> for DiscountPercentage {
     type Error = anyhow::Error;
-    fn try_from(value: f64) -> Result<Self, Self::Error> {
+    fn try_from(value: f32) -> Result<Self, Self::Error> {
         DiscountPercentage::new(value).context("")
     }
 }
@@ -168,7 +168,7 @@ mod discount_percentage_tests {
     #[case(101.0, false)]
     #[case(0.0, true)]
     #[case(20.0, true)]
-    fn test_discount_percentage(#[case] input: f64, #[case] valid: bool) {
+    fn test_discount_percentage(#[case] input: f32, #[case] valid: bool) {
         let p = DiscountPercentage::new(input);
         if valid {
             assert_that!(p).is_ok();
